--- src/app/demo/components/pages/empty/emptydemo.component.ts
@@ -1,6 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    templateUrl: './emptydemo.component.html'
-})
-export class EmptyDemoComponent { }
